refactor: cluster tests
